last_progress

+++ database/migrations/2024_08_27_080315_create_laporans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('laporans', function (Blueprint $table) {
             $table->id();
             $table->string('judul');
-            $table->integer('dana_realisasi');
+            $table->bigInteger('dana_realisasi');
             $table->date('tgl_realisasi');
             $table->enum('status', ['ditolak', 'diterima', 'revisi', 'draf'])->default('draf');
             $table->text("tanggapan")->nullable();
